refactor(ConfrimIntegrationEFormExtensions): move options object to own file, add sub folder for specifc vof
